requirements.txt
blinker            1.7.0
certifi            2023.11.17
charset-normalizer 3.3.2
click              8.1.7
colorama           0.4.6
Flask              3.0.0
idna               3.6
itsdangerous       2.1.2
Jinja2             3.1.2
MarkupSafe         2.1.3
pip                21.2.3
python-dotenv      1.0.0
requests           2.31.0
setuptools         57.4.0
urllib3            2.1.0
Werkzeug           3.0.1